% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddAttackSystem.R
\name{AddAttackSystem}
\alias{AddAttackSystem}
\title{Add Attack System}
\usage{
AddAttackSystem(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same object, with a new column (attack_system)
}
\description{
Creates an additional column indicating whether the attack is a high ball (Out of System) or not (In System)
}
