% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCover.R
\name{AddCover}
\alias{AddCover}
\title{Add Cover}
\usage{
AddCover(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same object, with "Cover" listed as a skill rather than skill_subtype
}
\description{
Replaces the "Dig" skill with "Cover" when the dig is classified as a spike cover
}
