% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddOpponent.R
\name{AddOpponent}
\alias{AddOpponent}
\title{Add Opponent}
\usage{
AddOpponent(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
A dv_plays object or data frame with a new column showing the opponent
}
\description{
Adds the opponent to a DataVolley play-by-play file
}
