% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddPlayerLocation.R
\name{AddPlayerLocation}
\alias{AddPlayerLocation}
\title{Add Player Location}
\usage{
AddPlayerLocation(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same object, with two new columns (player_location, player_position)
}
\description{
Creates an additional set of two columns indicating the court location and presumed position of the player
}
