% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddPossessions.R
\name{AddPossessions}
\alias{AddPossessions}
\title{Add Possessions}
\usage{
AddPossessions(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same object, with three new columns:
 possession_start - whether the touch starts a new possession,
 possession_number - the number of the possession - Serve = 1, first ball = 2, then incremented
 contact_number - the number of the contact within the possession

 Contact number 4 is common (usually corresponding to a block touch), contact numbers above 4 are rare but do occasionally happen
}
\description{
Adds the possession number and contact number within possession
}
