% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddPostAttackTouches.R
\name{AddPostAttackTouches}
\alias{AddPostAttackTouches}
\title{Add Post-Attack Touches}
\usage{
AddPostAttackTouches(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same object, with two more columns (attack_block, attack_dig).
For attack skills, the columns indicate the result of the block attempt and dig/cover attempt.
For other skills, the columns indicate whether the previous attack was block-touched and/or dug/covered.
}
\description{
Adds information about whether the current (previous) attack was touched on the block and/or dig attempt
}
