% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddRotations.R
\name{AddRotations}
\alias{AddRotations}
\title{Add Rotations}
\usage{
AddRotations(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
The same object, with four new columns (home_rotation, visiting_rotation, team_rotation, opponent_rotation)
}
\description{
Creates an additional set of four columns indicating the rotation numbers for the home team, visiting team, team performing the skill, and opposing team
}
