% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttackSummary.R
\name{AttackSummary}
\alias{AttackSummary}
\title{Attack Summary}
\usage{
AttackSummary(plays, ...)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}

\item{...}{Variables to group by.}
}
\value{
A grouped data frame
}
\description{
Summarizes all attacks in a DataVolley play-by-play file
}
