% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockSummary.R
\name{BlockSummary}
\alias{BlockSummary}
\title{Block Summary}
\usage{
BlockSummary(plays, ...)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}

\item{...}{Variables to group by.}
}
\value{
A grouped data frame
}
\description{
Summarizes all blocks in a DataVolley play-by-play file
}
