% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefenseSummary.R
\name{DigSummary}
\alias{DigSummary}
\title{Dig Summary}
\usage{
DigSummary(plays, ...)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}

\item{...}{Variables to group by.}
}
\value{
A grouped data frame
}
\description{
Summarizes all non-cover digs in a DataVolley play-by-play file
}
