% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorSummary.R
\name{ErrorSummary}
\alias{ErrorSummary}
\title{Error Summary}
\usage{
ErrorSummary(plays, ...)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}

\item{...}{Variables to group by.}
}
\value{
A grouped data frame
}
\description{
Summarizes all errors in a DataVolley play-by-play file
}
