% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstBallAttacks.R
\name{FirstBallAttacks}
\alias{FirstBallAttacks}
\title{FirstBallAttacks}
\usage{
FirstBallAttacks(plays)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}
}
\value{
A grouped data frame
}
\description{
Joins serve receptions in a DataVolley play-by-play file with the resulting attack (if any)
}
