% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReceptionSummary.R
\name{ReceptionSummary}
\alias{ReceptionSummary}
\title{Reception Summary}
\usage{
ReceptionSummary(plays, ...)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}

\item{...}{Variables to group by.}
}
\value{
A grouped data frame
}
\description{
Summarizes all serve receptions in a DataVolley play-by-play file
}
