% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServeSummary.R
\name{ServeSummary}
\alias{ServeSummary}
\title{Serve Summary}
\usage{
ServeSummary(plays, ...)
}
\arguments{
\item{plays}{a dv_plays object or data frame containing play-by-play data}

\item{...}{Variables to group by.}
}
\value{
A grouped data frame
}
\description{
Summarizes all serves in a DataVolley play-by-play file
}
